requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
email-validator